requirements.txt
watchdog==3.0.0
fastapi==0.104.1
uvicorn==0.23.2
anthropic==0.5.0
openai==1.3.0
pydantic==2.4.2
python-dotenv==1.0.0 